Added User Credibility Score
Alongside the rating for each restaurant, we include an overall
dependability score for all reviews submitted for that establishment.
The score is calculated by examining each user who has submitted a
review for that restaurant, comparing each of their ratings to the
average rating of each establishment. The difference between the scores
is summed up and divided by the number of reviews they have written.
Then the average value is averaged again among all reviewers for this
restaurant, giving the final score.